% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{annotation_position}
\alias{annotation_position}
\title{Create position-based annotations for A3}
\usage{
annotation_position(x, type = "")
}
\arguments{
\item{x}{Integer vector: Positions of the annotation (1-based indexing).}

\item{type}{Optional character scalar: Annotation type}
}
\value{
A3Position object
}
\description{
Creates an \code{A3Position} object for site, PTM, or processing annotations with
optional types.
}
\author{
EDG
}
