% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{annotation_range}
\alias{annotation_range}
\title{Create range-based annotations for A3}
\usage{
annotation_range(x, type = "")
}
\arguments{
\item{x}{Integer matrix with 2 columns corresponding to start and end positions of the annotation (1-based indexing).}

\item{type}{Optional character scalar: Annotation type}
}
\value{
A3Range object
}
\description{
Creates an \code{A3Range} object for region, PTM, or processing annotations with
optional types.
}
\author{
EDG
}
