% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{annotation_variant}
\alias{annotation_variant}
\title{Create variant annotations for A3}
\usage{
annotation_variant(x, info = list())
}
\arguments{
\item{x}{Integer scalar: Position of the variant (1-based indexing).}

\item{info}{Named list: Additional information about the variant (e.g. reference and alternate
amino acids, variant type, etc.)}
}
\value{
A3Variant object
}
\description{
Creates an \code{A3Variant} object for variant annotations with position and info.
}
\author{
EDG
}
