% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{create_A3}
\alias{create_A3}
\title{Create an A3 object from sequence, annotations, and metadata}
\usage{
create_A3(
  sequence,
  site = list(),
  region = list(),
  ptm = list(),
  processing = list(),
  variant = list(),
  uniprot_id = "",
  description = "",
  reference = "",
  organism = ""
)
}
\arguments{
\item{sequence}{Character: Amino acid sequence string.}

\item{site}{Named list of site annotations}

\item{region}{Named list of region annotations}

\item{ptm}{Named list of PTM annotations}

\item{processing}{Named list of processing annotations}

\item{variant}{Named list of variant annotations}

\item{uniprot_id}{Character: UniProt ID for metadata}

\item{description}{Character: Protein description for metadata}

\item{reference}{Character: Reference for metadata}

\item{organism}{Character: Organism name for metadata}
}
\value{
A3 object
}
\description{
Create an A3 object from sequence, annotations, and metadata
}
\author{
EDG
}
