% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene2sequence.R
\name{gene2sequence}
\alias{gene2sequence}
\title{Get the coding sequence of a gene}
\usage{
gene2sequence(
  gene,
  organism = "hsapiens",
  biomart = "ensembl",
  host = "https://www.ensembl.org",
  verbosity = 1L
)
}
\arguments{
\item{gene}{Character vector: One or more HGNC gene symbols.}

\item{organism}{Character scalar: Organism short name (Ensembl convention,
e.g. \code{"hsapiens"}).}

\item{biomart}{Character scalar: BioMart name.}

\item{host}{Character scalar: Host address.}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
data.frame with columns "gene", "ensembl_transcript_id" and "sequence".
}
\description{
Get the coding sequence of a gene
}
\examples{
# Requires internet connection and fetches data from Ensembl using biomaRt.
\dontrun{
  mapt_seq <- gene2sequence("MAPT")
}
}
\author{
EDG
}
