\docType{data}
\name{housing}
\alias{housing}
\title{Boston Housing Prices}
\description{
  A dataset containing the prices of houses in the Boston
  region and a number of features. The dataset and the
  following description are provided by UCI Machine
  Learning Repository
  (\url{http://archive.ics.uci.edu/ml/datasets/Housing}).
}
\details{
  \itemize{ \item CRIME: per capita crime rate by town
  \item ZN: proportion of residential land zoned for lots
  over 25,000 sq.ft.  \item INDUS: proportion of non-retail
  business acres per town \item CHAS: Charles River dummy
  variable (= 1 if tract bounds river; 0 otherwise) \item
  NOX: nitric oxides concentration (parts per 10 million)
  \item ROOMS: average number of rooms per dwelling \item
  AGE: proportion of owner-occupied units built prior to
  1940 \item DISTANCE: weighted distances to five Boston
  employment centres \item HIGHWAY: index of accessibility
  to radial highways \item TAX: full-value property-tax
  rate per $10,000 \item PUPIL.TEACHER: pupil-teacher ratio
  by town \item MINORITY: 1000(Bk - 0.63)^2 where Bk is the
  proportion of blacks by town \item LSTAT: % lower status
  of the population \item MED.VALUE: Median value of
  owner-occupied homes in $1000's }
}
\references{
  Frank, A. & Asuncion, A. (2010). UCI Machine Learning
  Repository [http://archive.ics.uci.edu/ml]. Irvine, CA:
  University of California, School of Information and
  Computer Science.

  Harrison, D. and Rubinfeld, D.L. Hedonic prices and the
  demand for clean air, J. Environ. Economics & Management,
  vol.5, 81-102, 1978.
}
\keyword{datasets}

