\docType{data}
\name{multifunctionality}
\alias{multifunctionality}
\title{Ecosystem Multifunctionality}
\description{
  This dataset relates multifunctionality to a number of
  different biotic and abiotic features in a global survey
  of drylands. The dataset was obtained from
  (\url{http://www.sciencemag.org/content/335/6065/214/suppl/DC1}).
  The dataset contains the features listed below.
}
\details{
  \itemize{ \item ELE: Elevation of the site \item LAT &
  LONG: Location of the site \item SLO: Site slope \item
  SAC: Soil sand content \item PCA_C1, PCA_C2, PCA_C3,
  PCA_C4: Principal components of a large set of primarily
  climatic features \item SR: Species richeness \item MUL:
  Multifunctionality }
}
\references{
  Maestre, F. T., Quero, J. L., Gotelli, N. J., Escudero,
  A., Ochoa, V., Delgado-Baquerizo, M., et al. (2012).
  Plant Species Richness and Ecosystem Multifunctionality
  in Global Drylands. Science, 335(6065), 214-218.
  doi:10.1126/science.1215442
}
\keyword{datasets}

