% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A3.R
\name{plot.A3}
\alias{plot.A3}
\title{Plot A3 Results}
\usage{
\method{plot}{A3}(x, ...)
}
\arguments{
\item{x}{an A3 object.}

\item{...}{additional options provided to \code{\link{plotPredictions}}, \code{\link{plotSlopes}} and \code{\link{plot}} functions.}
}
\description{
Plots an 'A3' object results. Displays predicted versus observed values for each observation along with the distribution of slopes measured for each feature.
}
\examples{
\donttest{
 data(housing)
 res <- a3.lm(MED.VALUE ~ NOX + ROOMS + AGE + HIGHWAY + PUPIL.TEACHER, housing, p.acc = NULL)
 plot(res)
 }
}

