% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{read_A3json}
\alias{read_A3json}
\title{Read \code{A3} object from JSON file}
\usage{
read_A3json(filepath, verbosity = 1L)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}

\item{verbosity}{Integer: if greater than 0, print messages.}
}
\value{
\code{A3} object.
}
\description{
Read \code{A3} object from JSON file
}
\examples{
a3 <- create_A3("MAEPRQEFEVMEDHAGTYGLGDRK", uniprot_id = "P10636")
tmp <- tempfile(fileext = ".json")
write_A3json(a3, tmp)
a3_read <- read_A3json(tmp, verbosity = 0L)
unlink(tmp)
}
\author{
EDG
}
