% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{write_A3json}
\alias{write_A3json}
\title{Write \code{A3} object to JSON file}
\usage{
write_A3json(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{A3} object.}

\item{filepath}{Character: Path to save JSON file.}

\item{overwrite}{Logical: If TRUE, overwrite existing file.}
}
\value{
Invisible \code{x}. Writes JSON file as side effect.
}
\description{
Write \code{A3} object to JSON file
}
\examples{
a3 <- create_A3("MAEPRQEFEVMEDHAGTYGLGDRK", uniprot_id = "P10636")
tmp <- tempfile(fileext = ".json")
write_A3json(a3, tmp)
a3_read <- read_A3json(tmp)
unlink(tmp)
}
\author{
EDG
}
