% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{concat}
\alias{concat}
\title{Concatenate character vector to single string for sequence input}
\usage{
concat(x)
}
\arguments{
\item{x}{Character vector to concatenate}
}
\value{
Single character string
}
\description{
Concatenate character vector to single string for sequence input
}
\examples{
# Concatenate a character vector of single-letter amino acids into a sequence string
concat(c("M", "A", "E", "P", "R"))
}
\author{
EDG
}
