% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A3.data.R
\docType{data}
\name{housing}
\alias{housing}
\title{Boston Housing Prices}
\usage{
data(housing)
}
\description{
A dataset containing the prices of houses in the Boston region and a number of features.
The dataset and the following description is based on that provided by UCI Machine Learning Repository (\url{http://archive.ics.uci.edu/ml/datasets/Housing}).
}
\details{
\itemize{
  \item CRIME: Per capita crime rate by town
  \item ZN: Proportion of residential land zoned for lots over 25,000 sq.ft.
  \item INDUS: Proportion of non-retail business acres per town
  \item CHAS: Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)
  \item NOX: Nitrogen oxides pollutant concentration (parts per 10 million)
  \item ROOMS: Average number of rooms per dwelling
  \item AGE: Proportion of owner-occupied units built prior to 1940
  \item DISTANCE: Weighted distances to five Boston employment centres
  \item HIGHWAY: Index of accessibility to radial highways
  \item TAX: Full-value property-tax rate per ten thousand dollar
  \item PUPIL.TEACHER: Pupil-teacher ratio by town
  \item MINORITY: 1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town
  \item LSTAT: Percent lower status of the population
  \item MED.VALUE: Median value of owner-occupied homes in thousands of dollars
}
}
\references{
Frank, A. & Asuncion, A. (2010). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.

Harrison, D. and Rubinfeld, D.L. Hedonic prices and the demand for clean air, J. Environ. Economics & Management, vol.5, 81-102, 1978.
}
\keyword{datasets}

