% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{read_A3json}
\alias{read_A3json}
\title{Read \code{A3} object from JSON file}
\usage{
read_A3json(filepath, verbosity = 0L)
}
\arguments{
\item{filepath}{Character: Path to JSON file.}

\item{verbosity}{Integer: if greater than 0, print messages.}
}
\value{
\code{A3} object.
}
\description{
Read \code{A3} object from JSON file
}
\examples{
\dontrun{
mapt <- uniprot_to_A3("P10636")
write_A3json(mapt, "P10636_A3.json")
mapt2 <- read_A3json("P10636_A3.json")
}
}
\author{
EDG
}
