% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3-package.R
\docType{package}
\name{rtemis.a3-package}
\alias{rtemis.a3}
\alias{rtemis.a3-package}
\title{rtemis.a3: Annotated Amino Acid format}
\description{
Annotated Amino Acid format utilities
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rtemis.a3.rtemis.org}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID})

}
