% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a3.R
\name{write_A3json}
\alias{write_A3json}
\title{Write \code{A3} object to JSON file}
\usage{
write_A3json(x, filepath, overwrite = FALSE)
}
\arguments{
\item{x}{\code{A3} object.}

\item{filepath}{Character: Path to save JSON file.}

\item{overwrite}{Logical: If TRUE, overwrite existing file.}
}
\value{
Invisible \code{x}. Writes JSON file as side effect.
}
\description{
Write \code{A3} object to JSON file
}
\examples{
\dontrun{
mapt <- uniprot_to_A3("P10636")
write_A3json(mapt, "P10636_A3.json")
}
}
\author{
EDG
}
