% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/A3.R
\name{xtable.A3}
\alias{xtable.A3}
\title{Nicely Formatted Fit Results}
\usage{
\method{xtable}{A3}(x, ...)
}
\arguments{
\item{x}{an A3 object.}

\item{...}{additional arguments passed to the \code{\link{print.xtable}} function.}
}
\description{
Creates a LaTeX table of results. Depends on the \pkg{xtable} package.
}
\examples{
x <- a3.lm(rating ~ ., attitude, p.acc = NULL)
 xtable(x)
}

